🔧 chore start splitting between CLI and lib
