+++ packages/cozy-jobs-cli/bin/cozy-run-shell
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+
+require(require('path').join(__dirname, '../src', 'shell'))
