+++ packages/cozy-jobs-cli/package.json
@@ -1,0 +1,29 @@
+{
+  "name": "cozy-jobs-cli",
+  "version": "1.0.0",
+  "description": "Manage cozy jobs",
+  "main": "index.js",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/cozy/cozy-konnector-libs.git"
+  },
+  "license": "MIT",
+  "bin": {
+    "cozy-run-shell": "bin/cozy-run-shell",
+    "cozy-run-standalone": "bin/cozy-run-standalone",
+    "cozy-run-dev": "bin/cozy-run-dev"
+  },
+  "homepage": "https://github.com/cozy/cozy-konnector-libs#readme",
+  "dependencies": {
+    "cheerio": "^1.0.0-rc.2",
+    "commander": "^2.12.2",
+    "opn": "^5.1.0"
+  },
+  "scripts": {
+    "lint": "standard --fix src/*"
+  },
+  "devDependencies": {
+    "cross-env": "^5.1.3",
+    "lerna": "^2.4.0"
+  }
+}
