+++ packages/cozy-jobs-cli/src/cozy-authenticate.js
