+++ packages/cozy-jobs-cli/src/init-dev-account.js
