+++ packages/cozy-jobs-cli/src/init-konnector-config.js
