+++ packages/cozy-jobs-cli/src/shell.js
