+++ packages/cozy-jobs-cli/src/standalone.js
@@ -7,7 +7,7 @@ process.env.COZY_FIELDS = JSON.stringify({
   folder_to_save: '.'
 })
 
-const config = require('../helpers/init-konnector-config')()
+const config = require('cozy-konnector-libs/helpers/init-konnector-config')()
 process.env.COZY_URL = config.COZY_URL
 
 const filename = process.argv[2] || 'index.js'
