--- packages/cozy-konnector-libs/bin/cozy-konnector-standalone
@@ -1,3 +1,0 @@
-#!/usr/bin/env node
-
-require(require('path').join(__dirname, '../commands', 'standalone'))
